added confirmation dialog when user wants to delete his account

+++ frontend/src/app/account-manager/account-manager.component.ts
@@ -7,9 +7,11 @@ import {
   Form,
 } from '@angular/forms';
 import { Router } from '@angular/router';
+import { MatDialog } from '@angular/material/dialog';
 
 import { LocalStorageService } from '../services/localstorage.service';
 import { AccountService } from '../services/account.service';
+import { ConfirmJobsiteDeleteComponent } from '../confirm-jobsite-delete/confirm-jobsite-delete.component';
 
 @Component({
   selector: 'app-account-manager',
@@ -21,21 +23,28 @@ export class AccountManagerComponent implements OnInit {
     private formBuilder: FormBuilder,
     private localStorageService: LocalStorageService,
     private accountService: AccountService,
-    private router: Router
+    private router: Router,
+    public dialog: MatDialog
   ) {}
 
   ngOnInit(): void {}
 
   onDeleteAccount() {
-    this.accountService.deleteUserAccount().subscribe(
-      (res) => {
-        this.localStorageService.clear();
-        this.router.navigate(['']);
-      },
-      (err) => {
-        console.log(err);
+    const dialog = this.dialog.open(ConfirmJobsiteDeleteComponent);
+
+    dialog.afterClosed().subscribe((result) => {
+      if (result) {
+        this.accountService.deleteUserAccount().subscribe(
+          (res) => {
+            this.localStorageService.clear();
+            this.router.navigate(['']);
+          },
+          (err) => {
+            console.log(err);
+          }
+        );
       }
-    );
+    });
   }
 
   onBack() {
